Create a R project with full content skeleton

Create a R project with full content skeleton

create_r_project(
  project_path,
  type = c("package", "golem", "book"),
  name_licence,
  type_licence
)

Arguments

project_path

project_path

type

type

name_licence

name for the licence

type_licence

type of the licence, should be a function like usethis::use_mit_licence

Value

Side Effect: Transform project as package or book:

With package, golem or book skeleton

git necessary files

pkgdown or book template

GitLab CI

Examples

### R package skeleton
withr::with_tempdir({
  project_path <- getwd()
  create_r_project(
    project_path,
    type = c("package"),
    name_licence = "Bob",
    type_licence = usethis::use_proprietary_license
  )
})
#> ✔ Setting active project to '/tmp/RtmppzqM1o/file16d4205d7f22'
#> ✔ Creating 'R/'
#> ✔ Writing a sentinel file '.here'
#> • Build robust paths within your project via `here::here()`
#> • Learn more at <https://here.r-lib.org>
#> ✔ Setting active project to '<no active project>'
#> ✔ Setting active project to '/tmp/RtmppzqM1o/file16d4205d7f22'
#> ✔ Setting active project to '/tmp/RtmppzqM1o/file16d4205d7f22'
#> ✔ Writing 'DESCRIPTION'
#> Package: file16d4205d7f22
#> Title: What the Package Does (One Line, Title Case)
#> Version: 0.0.0.9000
#> Authors@R (parsed):
#>     * First Last <first.last@example.com> [aut, cre] (YOUR-ORCID-ID)
#> Description: What the package does (one paragraph).
#> License: `use_mit_license()`, `use_gpl3_license()` or friends to
#>     pick a license
#> Encoding: UTF-8
#> Roxygen: list(markdown = TRUE)
#> RoxygenNote: 7.2.3
#> ✔ Writing 'NAMESPACE'
#> ✔ Setting active project to '/tmp/RtmppzqM1o/file16d4205d7f22'
#> Saving attachment parameters to yaml config file
#> Updating file16d4205d7f22 documentation
#> ℹ Loading file16d4205d7f22
#> ℹ Loading file16d4205d7f22
#> ✔ Setting active project to '<no active project>'
#> ✔ Setting active project to '/tmp/RtmppzqM1o/file16d4205d7f22'
#> ✔ Adding 'file LICENSE' to License
#> ✔ Writing 'LICENSE'
#> ✔ Writing 'NEWS.md'
#> ✔ Setting active project to '<no active project>'
#> Kit project created
#> ✔ Setting active project to '/tmp/RtmppzqM1o/file16d4205d7f22'
#> ✔ Adding 'testthat' to Suggests field in DESCRIPTION
#> ✔ Adding '3' to Config/testthat/edition
#> ✔ Creating 'tests/testthat/'
#> ✔ Writing 'tests/testthat.R'
#> • Call `use_test()` to initialize a basic test file and open it for editing.
#> ✔ Writing 'tests/testthat/test-start.R'
#> • Edit 'tests/testthat/test-start.R'
#> ✔ Setting active project to '<no active project>'
#> Kit package created
#> [1] "/tmp/RtmppzqM1o/file16d4205d7f22"

### Shiny app
withr::with_tempdir({
  project_path <- getwd()
  create_r_project(
    project_path,
    type = c("golem"),
    name_licence = "Bob",
    type_licence = usethis::use_proprietary_license
  )
})
#> ✔ Setting active project to '/tmp/RtmppzqM1o/file16d4584f25f0'
#> ✔ Creating 'R/'
#> ✔ Writing a sentinel file '.here'
#> • Build robust paths within your project via `here::here()`
#> • Learn more at <https://here.r-lib.org>
#> ✔ Setting active project to '<no active project>'
#> ✔ Setting active project to '/tmp/RtmppzqM1o/file16d4584f25f0'
#> ── Checking package name ───────────────────────────────────────────────────────
#> ✔ Valid package name
#> • Overwriting existing project.
#> ── Copying package skeleton ────────────────────────────────────────────────────
#> ✔ Copied app skeleton
#> ── Running project hook function ───────────────────────────────────────────────
#> ✔ All set
#> ── Done ────────────────────────────────────────────────────────────────────────
#> A new golem named file16d4584f25f0 was created at /tmp/RtmppzqM1o/file16d4584f25f0 .
#> To continue working on your app, start editing the 01_start.R file.
#> We noticed that some dev dependencies are not installed.
#> You can install them with `golem::install_dev_deps()`.
#> Saving attachment parameters to yaml config file
#> Updating file16d4584f25f0 documentation
#> Setting `RoxygenNote` to "7.2.3"
#> ℹ Loading file16d4584f25f0
#> Writing NAMESPACE
#> Writing NAMESPACE
#> Writing run_app.Rd
#> ℹ Loading file16d4584f25f0
#> ✔ Setting active project to '<no active project>'
#> ✔ Setting active project to '/tmp/RtmppzqM1o/file16d4584f25f0'
#> ✔ Adding 'file LICENSE' to License
#> ✔ Writing 'LICENSE'
#> ✔ Writing 'NEWS.md'
#> ✔ Setting active project to '<no active project>'
#> Kit project created
#> ✔ Setting active project to '/tmp/RtmppzqM1o/file16d4584f25f0'
#> ✔ Adding 'testthat' to Suggests field in DESCRIPTION
#> ✔ Adding '3' to Config/testthat/edition
#> ✔ Creating 'tests/testthat/'
#> ✔ Writing 'tests/testthat.R'
#> • Call `use_test()` to initialize a basic test file and open it for editing.
#> ✔ Writing 'tests/testthat/test-start.R'
#> • Edit 'tests/testthat/test-start.R'
#> ✔ Setting active project to '<no active project>'
#> Kit package created
#> [1] "/tmp/RtmppzqM1o/file16d4584f25f0"

### Bookdown
withr::with_tempdir({
  project_path <- getwd()
  create_r_project(
    project_path,
    type = c("book"),
    name_licence = "Bob",
    type_licence = usethis::use_proprietary_license
  )
})
#> [-] 5 package(s) removed: dplyr, ggplot2, pagedown, tidyr, pagedown.
#> book created
#> You need to follow 'dev_history_book.Rmd' after that for the CI to create the pdf_book.
#> ✔ Setting active project to '/tmp/RtmppzqM1o/file16d418ca6e2f'
#> ✔ Writing 'LICENSE'
#> ✔ Writing 'NEWS.md'
#> ✔ Setting active project to '<no active project>'
#> Kit project created
#> [1] "/tmp/RtmppzqM1o/file16d418ca6e2f"